% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confbounds.R
\name{confbounds}
\alias{confbounds}
\title{Add Confidence Bounds to Linear Model Plots}
\usage{
confbounds(model, xbounds = range(model$model[, 2]), level = 0.95, ...)
}
\arguments{
\item{model}{an object containing the results returned by a
model fitting function (e.g., lm or glm).}

\item{xbounds}{left-right limits on the lines to be drawn.}

\item{level}{float (0,1); the confidence level, defaults to 0.95.}
}
\description{
Draws in confidence line segments based on model objects.
}
\examples{
confbars(mod1, lty=2, col='red')
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\keyword{'linear}
\keyword{model'}
\keyword{plotting,}

