% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roller.R
\name{roller}
\alias{roller}
\title{Die Roller}
\usage{
roller(sides_vec, rolls_vec)
}
\arguments{
\item{sides_vec}{vector where each element is a number of sides of a die}

\item{rolls_vec}{vector of times to roll each corresponding die}
}
\value{
returns a list of vectors containing roll results for each die shape
}
\description{
Roll any number of dice any number of times apiece
}
\examples{
roller(c(6,12,20), 1:3)

$`6-sided`
[1] 5

$`12-sided`
[1] 12 11

$`20-sided`
[1]  3 18 20
}
\author{
Mike Vlah, \email{vlahm13@gmail.com}
}
\keyword{probability}

